refactor: rename folder as test-endpoint-table
