+++ dottie-app/src/api/index.ts
@@ -2,12 +2,12 @@ import apiClient from './core/apiClient';
 import { isSuccess, isClientError, isServerError } from './core/apiClient';
 import { checkDbConnection, fetchUserData } from './core/db';
 
-import authApi from './auth';
+import { authApi } from './auth';
 import { type User, type LoginInput, type SignupInput } from './auth/types';
 
-import assessmentApi, { type Assessment } from './assessment';
-import messageApi, { type ApiMessage } from './message';
-import userApi, { type UserProfile } from './user';
+import { assessmentApi, type Assessment } from './assessment';
+import { messageApi, type ApiMessage } from './message';
+import { userApi, type UserProfile } from './user';
 
 // Export all API modules
 export {
