+++ dottie-app/src/pages/test_page/test-endpoint-table/chat/ChatEndpoints.tsx
